First draft, uploading a copy of TwentyTwelve.
In this draft I'm committing a copy of TwentyTwelve. I haven't started
working on the final theme itself yet, but I've got the logo complete. I
was going to do the theme without a PSD (I already know how the site is
going to look), but I've decided I WILL create a PSD file for the theme
before I work on the template itself. For now, here's a copy of
TwentyTwelve to get started, and I'll update it later.