Added styles for the homepage.
I finally went in and updated the styles to get this thing done. I set
all the styles for the layout, cut Gumby (it wouldn't work well with
Less), and changed this theme to jrBlog. The theme will work
out-of-the-box once complete, but for now I'm focusing on getting it
done specifically for my personal website. Once the site is online I'll
continue to expand the theme further in the future.