Finished Footer and Sidebar
I finished the footer and sidebar, and redid the main content area to
look more like my original design. I also replaced the footer with a
texture and gradient overlay. It doesn't look as good, but I can now
make the footer taller if I need to. I still have to add the footer and
header sidebar widgets, but other than that and some additional testing
on other pages its basically done.