Added Support for Child Themes
I've had this up for a while now, I made some improvements while using
this theme as a base for new themes I've created. This update includes
additional support for child themes and also removes the base colors of
the theme to be grays. The only thing left I know I need to work on here
is responsive design, which I still need to finish up.